+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/datenlord/aligned-utils/compare/v1.0.0...HEAD
 
+### Added
+
++ impl `Clone` for `AlignedBytes`
+
 ## [1.0.0] - 2021-01-12
 
 [1.0.0]: https://github.com/datenlord/aligned-utils/compare/v0.2.0...v1.0.0